+++ src/rules/common/fileWarnings.js
@@ -6,13 +6,16 @@ import fs from 'fs';
 
 import getMessageLogger from '../getMessageLogger';
 
-export default (file, { logType, msg } = {}) => {
+export default (file, { logType, msg, ignoreRegex } = {}) => {
   if (!file) {
     warn('`commonFileWarnings`: missing "file" parameter');
   } else {
     const contents = fs.readFileSync(file).toString();
-    const warnings = contents.match(/^.*\bwarning( |:).*$/gim);
-    if (warnings) {
+    let warnings = contents.match(/^.*\bwarning( |:).*$/gim);
+    if (ignoreRegex) {
+      warnings = warnings.filter(warning => !warning.match(ignoreRegex));
+    }
+    if (warnings && warnings.length) {
       const log = getMessageLogger(logType);
       const warningsStr = warnings.map(line => `- ${line}`).join('\n');
       const introMsg =